# application/services/custom_services/tickets/src/models/entitymodels/ticket.ts
export interface ticket 
{
   created_date: { type: Date },
   created_by: { type: String },
   last_modified_by: { type: String },
   last_modified_date: { type: Date },
   name: String,
   description: String
}